+++ packages/ui/src/public-api.ts
@@ -59,6 +59,7 @@ export * from './lib/tabs/tab.component';
 export * from './lib/tag/tag.component';
 export * from './lib/textarea/textarea.component';
 export * from './lib/time-picker/time-picker.component';
+export * from './lib/timeline/timeline.component';
 export * from './lib/toast/toast.component';
 export * from './lib/toast/toast.service';
 export * from './lib/tooltip/tooltip.directive';